refactor(routes): inline top-level page components into route files
Moves each route's top-level page component (e.g. AdminLayout,
ArtistBulkEditor, EditionLayout, ListSchedule) directly into its
corresponding src/routes/ file, while leaving sub-component folders in
place under src/pages/ and fixing import paths accordingly.
