--- src/pages/EditionView/tabs/ScheduleTab/TimelineTab.tsx
@@ -1,5 +1,0 @@
-import { Timeline } from "./horizontal/Timeline";
-
-export function ScheduleTabTimeline() {
-  return <Timeline />;
-}
