--- src/pages/admin/festivals/FestivalSets.tsx
@@ -1,5 +1,0 @@
-import { SetManagement } from "./SetsManagement/SetManagement";
-
-export default function FestivalSets() {
-  return <SetManagement />;
-}
