--- src/pages/admin/festivals/FestivalStages.tsx
@@ -1,5 +1,0 @@
-import { StageManagement } from "./StageManagement";
-
-export default function FestivalStages() {
-  return <StageManagement />;
-}
