+++ src/routes/admin/festivals/$festivalSlug/editions/$editionSlug/sets.tsx
@@ -1,8 +1,12 @@
 import { createFileRoute } from "@tanstack/react-router";
-import FestivalSets from "@/pages/admin/festivals/FestivalSets";
+import { SetManagement } from "@/pages/admin/festivals/SetsManagement/SetManagement";
 
 export const Route = createFileRoute(
   "/admin/festivals/$festivalSlug/editions/$editionSlug/sets",
 )({
   component: FestivalSets,
 });
+
+function FestivalSets() {
+  return <SetManagement />;
+}
