+++ src/routes/admin/festivals/$festivalSlug/editions/$editionSlug/stages.tsx
@@ -1,8 +1,12 @@
 import { createFileRoute } from "@tanstack/react-router";
-import FestivalStages from "@/pages/admin/festivals/FestivalStages";
+import { StageManagement } from "@/pages/admin/festivals/StageManagement";
 
 export const Route = createFileRoute(
   "/admin/festivals/$festivalSlug/editions/$editionSlug/stages",
 )({
   component: FestivalStages,
 });
+
+function FestivalStages() {
+  return <StageManagement />;
+}
